move celestial browser and active launches windows
