+++ src/modules/window/active_launches_window.h
